Move tests into the crate source
We don't really need integration tests that much. The API is documented
in rustdocs, so we can test all the stuff we need right in the crate
source code. Integration tests usually serve as example code, but ours
are not really that clean. Just read the API docs.
